Fix dedup URL params
The depedency, url-parse, will dedup repeated url params. For example
`?fop=1&foo2` and will keep the first one. This doesn't play nicely with rails
Rack params as dups may indicate a structural array. This change migrates
`url-parse` to the browser's own 'URLSearchParams' and `URL`.

With the change, we had to:
1. Make minimal changes to `url.ts` helpers. You'll notice that we have a FAKE_ORIGIN as a fallback for the
`URL` usage, but in every instances, we have no 't need the origin at all. Either we replace the origin with a blank
or we only a part of the url like pathname.
2. baseUrl is required, but existing testing in isolation has it as optional. The update to the tests now reflect
that requirement

For dev envs, we're also included corejs which helps with adding polyfills that
we can use in place for testing.